refactor(sync): extract inner classes from FlinkSqlWDS and add tutorial comments
- Extract CdcEventDeserializer to standalone class (Kafka SourceRecord → CDC Event)
- Extract CDBBatchSink to standalone class (batch JDBC upsert/delete sink)
- Move mapType, createSinkTableIfNotExists, validateIdentifier to MysqlUtil
- Add tutorial-oriented Javadoc and inline comments to all four files
- Update FlinkSqlWDSTest to reference MysqlUtil.mapType
